move pages to components folder and rename them, rename lobby step forstats for consistency
